[ctreemanifest] fix ambiguous call to string::erase()
Summary:
gcc 4.9 complains that erase(0) is ambiguous, and may be either string::erase(size_t)
or string::erase(iterator) (since iterator is "char*", and for historical
reasons 0 can be interpreted as a null char*).

Fix the code to explicitly indicate that it means the erase(size_t) version.

Test Plan: Confirmed that the code built successfully with gcc 4.9.

Reviewers: durham, mitrandir, ttung

Reviewed By: ttung

Subscribers: net-systems-diffs@, yogeshwer, mjpieters

Differential Revision: https://phabricator.intern.facebook.com/D3911252

Signature: t1:3911252:1474589551:688ecab4d59053dbdf7b48a062f248d8363e17f3

+++ ctreemanifest/py-treemanifest.cpp
@@ -345,7 +345,7 @@ static bool fileiter_popfinished(fileiter *iter) {
     if (found != std::string::npos) {
       iter->path.erase(found + 1);
     } else {
-      iter->path.erase(0);
+      iter->path.erase(size_t(0));
     }
   }
